requirements.txt
algoliasearch==2.1.0
analytics-python==1.2.9
appdirs==1.4.3
astroid==2.3.3
attrs==19.3.0
beautifulsoup4==4.8.2
black==19.10b0
boto3==1.10.45
botocore==1.13.45
cachetools==4.1.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.7
django-cors-headers==3.2.0
django-heroku==0.3.1
django-model-utils==4.0.0
django-redis-cache==2.1.0
django-rest-knox==4.1.0
django-rq==2.3.0
django-storages==1.8
djangorestframework==3.10.3
djangorestframework-recursive==0.1.2
docutils==0.15.2
google-api-core==1.17.0
google-api-python-client==1.8.0
google-auth==1.14.0
google-auth-httplib2==0.0.3
googleapis-common-protos==1.51.0
gunicorn==20.0.0
html-sanitizer==1.9.0
httplib2==0.17.2
idna==2.8
inflection==0.3.1
isort==4.3.21
itypes==1.1.0
Jinja2==2.11.1
jmespath==0.9.4
lazy-object-proxy==1.4.3
lxml==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.1
pathspec==0.6.0
protobuf==3.11.3
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.4
pylint-django==2.0.12
pylint-plugin-utils==0.6
pyparsing==2.4.6
python-dateutil==2.8.1
python-http-client==3.2.4
python-postmark==0.5.3
python3-openid==3.1.0
pytz==2019.3
rcssmin==1.0.6
redis==3.3.11
regex==2019.11.1
requests==2.22.0
requests-oauthlib==1.3.0
rq==1.2.2
rsa==4.0
ruamel.yaml==0.16.7
ruamel.yaml.clib==0.2.0
s3transfer==0.2.1
sendgrid==6.1.1
sentry-sdk==0.13.5
simplejson==3.17.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
tinycss2==1.0.2
toml==0.10.0
typed-ast==1.4.0
uritemplate==3.0.1
urllib3==1.25.7
webencodings==0.5.1
whitenoise==4.1.4
wrapt==1.11.2
